+++ include/controller/c/webots/nodes.h
@@ -84,6 +84,7 @@ typedef enum {
   // misc
   WB_NODE_BALL_JOINT,
   WB_NODE_BALL_JOINT_PARAMETERS,
+  WB_NODE_BASIC_HINGE_JOINT,
   WB_NODE_CHARGER,
   WB_NODE_CONTACT_PROPERTIES,
   WB_NODE_DAMPING,
