+++ src/webots/nodes/utils/WbConcreteNodeFactory.cpp
@@ -44,6 +44,7 @@
 #include "WbGroup.hpp"
 #include "WbGyro.hpp"
 #include "WbHinge2Joint.hpp"
+#include "WbHingeJoint.hpp"
 #include "WbHingeJointParameters.hpp"
 #include "WbImageTexture.hpp"
 #include "WbImmersionProperties.hpp"
@@ -119,6 +120,8 @@ WbNode *WbConcreteNodeFactory::createNode(const QString &modelName, WbTokenizer 
     return new WbBallJoint(tokenizer);
   if (modelName == "BallJointParameters")
     return new WbBallJointParameters(tokenizer);
+  if (modelName == "BasicHingeJoint")
+    return new WbBasicHingeJoint(tokenizer);
   if (modelName == "Box")
     return new WbBox(tokenizer);
   if (modelName == "Brake")
@@ -307,6 +310,8 @@ WbNode *WbConcreteNodeFactory::createCopy(const WbNode &original) {
     return new WbBallJoint(original);
   if (modelName == "BallJointParameters")
     return new WbBallJointParameters(original);
+  if (modelName == "BasicHingeJoint")
+    return new WbBasicHingeJoint(original);
   if (modelName == "Box")
     return new WbBox(original);
   if (modelName == "Brake")
